requirements.txt
timm==1.0.8 
openai==1.40.3 
lmdeploy[all]==0.5.3
langchain==0.2.13
langchain-chroma==0.1.2
langchain-community==0.2.12
langchain-core==0.2.30
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langchainhub==0.1.21
PyYaml
streamlit==1.36.0
streamlit-keyup==0.2.4
funasr==1.0.0
streamlit-audiorecorder==0.0.5

#tts
sortedcontainers
simplejson
